+++ types/affinda.d.ts
@@ -2072,19 +2072,19 @@ export declare type ResthookEvent = string;
 
 export declare interface ResthookSubscription {
     /** Resthook subscription's ID. */
-    id?: number;
+    id: number;
     /** The event name to subscribe to. */
-    event?: ResthookEvent;
+    event: ResthookEvent;
     /** URL of the resthook's receiver. */
-    targetUrl?: string;
+    targetUrl: string;
     /** Resthooks only fire for active subscriptions. */
-    active?: boolean;
+    active: boolean;
     /** Resthook subscriptions can be auto deactivated if the receiver continuously returns error status code over a period of time. */
-    autoDeactivated?: boolean;
+    autoDeactivated: boolean;
     /** The reason for the subscription being auto deactivated. May contains the error response that the receiver returned. */
-    autoDeactivateReason?: string;
+    autoDeactivateReason: string;
     /** Version of the resthook subscription. Determines the resthook body being fired. */
-    version?: ResthookSubscriptionVersion;
+    version: ResthookSubscriptionVersion;
 }
 
 export declare interface ResthookSubscriptionCreate {