+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add international_country_code to phone number details
 - Add customData field to JD search param/result serializers
 - rawText field to JobDescription Model
+- Required fields for resthook subscriptions
 
 ## [5.3.1] - 2023-03-30
 ### Changed